Fix scraping recent jobs
Fix scraping recent jobs by using date-time module to remove jobs with opening dates older than "19-05-2020". Before, there was a bug that excluded jobs with an opening date of "19-05-2020".

Refactor the check_before_extract function to implement the logic for the checking the dates of the jobs that was originally inside the extract_job_details function.