Feature job deadlines
Feature the jobs' deadlines by entering the jobs' URL and scraping for the expiry date. This date is then appended to the Excel worksheet in the new "Deadline" column.

However, this feature is known to make the program run slower as it needs to access each of the job's web page to scrape the deadlines. Also, the format of the deadline is different from the job's posted date as seen after creating the Excel workbook. Additionally, there maybe some expired jobs included in the worksheet.