Fix salary scraping and request errors
Fix a problem where unpaid and voluntary roles do not have a salary range given in the salary section but specifies that it's unpaid in the job description itself. These come up in the salary range column as "Unpaid".

Also, fix an issue where an connection error occurs when the making get requests.