Add competitve salary and company links
Add competitive salary as a new salary range. Jobs with competitive salaries are found by searching through the text of the job description page, similar to how unpaid salaries were found.

Also, add a new feature to scrape the company website link from the job description page and append it to the Excel worksheet in the "Company" column.