Add commission and equity salaries
Add new salaries types to the salary range. These include commission-based salaries and others that include equity. When commission and equity are given on top of a salary then another string is appended onto the "salary" the salary_range variable to specify this.

Also, improved the logic for finding unpaid salaries by scraping for jobs that have "no salary" specified in their description.

Changed the ungiven salary ranges from "Not specified" to "Unspecified salary".