Refactor removing jobs function
Refactor the remove_outdated_jobs() function to move the checking dates of the jobs to another new function called check_date(). This new function checks only the last job's date of the job list to check whether to keep searching before moving to the next search page. This helps to speed up the process and uses less memory by not iterating over a whole copy of the job list but only checking the last entered jobs at certain times.

Also, moved the remove_outdated_jobs() function outside the while loop in the search_for_jobs() function. This speeds up the process as before this function was called in the while loop which was unnecessary and inefficient.

Also, deleted the sleep wait inside search_for_jobs() as it was no longer needed due to the sleep now being called in the soup_creator() function.