Refactor web_scraper
Refactor how web_scraper.py scrapes each job posting. Scrape_job_details() is changed to scrape one job posting which is called by a new function scrape_job_details() to scrape the whole page and check it's posted date before appending it to excel.

This improves speeds as outdated jobs are no longer appended to job_list and then removed afterwards. Also, uses less memory as copies of the list and other variables are no longer necessary and uses less lines of code as check_date() and remove_outdated_jobs are no longer needed.

Also, refactor soup_creator() to not specify the value for max_retry in get_request() but instead let it be set when soup_creator() is called.