Add equity only and refactor scrape salaries
Add "Equity only" as a new salary range for jobs that only offer equity. This is similar to how commission-based jobs are sorted, either as a remuneration or the only form of payment.

Also, refactor how some of the salary ranges are scraped by using list  comprehension to use less lines of code.