Improve finding salary range
Improve finding the salary range by searching through the text of the job description. Some jobs don't state the salary range in the salary section of the job description but in the body of text instead. These salaries are found by looking for common characters like "£" and "000".

These scrapped strings are then reformatted to a similar format as the other salary ranges by removing and adding characters.