Improve salary range format
Improve the format of the salary ranges found in the body of text of the job description to be consistent with the other salary formats. For instance, the new logic formats the salary range into "{lower salary amount} - {upper salary amount} per year". In other words, "-" is added inbetween with spaces and "per year" is added to the end of the string.

However, there are a few bugs including not registering salaries written in "k" rather than "000" and renumerations that aren't paid yearly getting "per year"appended on the end.