Refactor salary range
Refactor how the salary range is scraped from the text of the job description by using less variables, replacing the lists with tuples to use less memory and combining some of the if conditions into a for loop. This helps to use less lines of code and speeds up the process.