Fix salary range IndexError and Add salary estimate
Fix the IndexError that occurred after looking ahead of the string when searching for the salary range. This is solved by removing it as a conditional and using an exception instead.

Also, add conditionals to check ahead of the salary range string to check what basis the salary is paid on (e.g per hour, per week, etc.).