Add salary_finder function
Add a new function called salary_finder that finds salaries in the text of the job description. The salary_finder takes two parameters, the soup of the job description webpage and the text tags to search through.

This new function fixs the issue where some salary ranges are written in the <li> tags as well as the <p> tags. To solve this the salary_finder is called twice to search through both tags and returns a string to be used for salary_range.