Add bonus to salary
Add bonuses as a new salary add-on to the salary_range variable. The program finds whether a job offers bonuses by scraping through the text of the job description page for any sentences that mentions the word "Bonus". Then, rules are applied to the sentences to test whether it means bonus skills/experiences or remuneration bonuses.

Refactor the operations on the text_list in the salary_finder() function to only lowercase the list once rather than for each conditional.