Add load workbook feature
Add a feature that loads the workbook from the excel directory if it already exists. If not, then it creates a new workbook.

To avoid the same jobs being appended to workbook again, the append_jobs_to_xl() functions checks whether each scraped job is already in the workbook before appending it. Once it finds a job that is already in the worksheet, then it stops appending jobs.

Also, rename some functions to not conflict with the openpyxl methods and to better describe their purpose.