Fix checks whether job's already appended
Fix a problem were checking whether the job is already in the worksheet did not work, causing duplicate job postings. This is caused by job postings mostly changing minor details meaning that the program recognises them as different jobs.

Known problems are that jobs that are updated with important details on workinstartups.com would not be updated in the worksheet. Also, newer jobs are appended at the bottom rather than the top of the worksheet.