Add sorted job list function
Add a new function called sort_job_list() to take all of the jobs from worksheet, sort them out based on the opening dates (earliest first) and then append the new list to the worksheet.

Known problems are that the jobs in the worksheet has to be deleted first before appending the new sorted job list. Also, the job_list is appended twice to the worksheet which is inefficient way of organising the job list. Once for adding new jobs, and the second time for the new order of the jobs.