Add function to get first job date
Add a new function called get_first_job_date() to retrieve the date of the first job in the excel worksheet. This is date is used to know when to stop scraping for jobs on workinstartups.com. This is used when there's a pre-existing workbook with jobs already in the table. It will scrape for jobs until the last date since the site was scraped before. Else, the program will create a new workbook and scrape jobs posted since a month ago.

Also, improved comments to better describe the functionality of the modules and functions.